creating PvtArgs, PvtViewModel classes
