+++ AndroidPvtLib/src/main/java/rs/arthu/androidpvt/lib/PvtActivity.kt
@@ -7,12 +7,22 @@ import android.os.Bundle
 import android.util.Log
 import android.view.MotionEvent
 import android.view.View
+import androidx.lifecycle.ViewModelProvider
 import rs.arthu.androidpvt.lib.databinding.ActivityPvtBinding
+
+const val PVT_RESULTS_KEY = "pvtResultsKey"
+internal const val STIMULUS_COUNT = "stimulusCount"
+internal const val MIN_INTERVAL = "minInterval"
+internal const val MAX_INTERVAL = "maxInterval"
+internal const val COUNTDOWN_TIME = "countdownTime"
+internal const val STIMULUS_TIMEOUT = "stimulusTimeout"
+internal const val POST_RESPONSE_DELAY = "postResponseDelay"
 
 class PvtActivity : AppCompatActivity() {
 
-    private var pvt: Pvt? = null
     private lateinit var binding: ActivityPvtBinding
+    private lateinit var viewModel: PvtViewModel
+    private lateinit var viewModelFactory: PvtViewModelFactory
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -22,65 +32,17 @@ class PvtActivity : AppCompatActivity() {
         setContentView(view)
 
         val intentExtras = intent.extras
-        val pvtArgs = getPvtArgsFromIntentExtras(intentExtras)
+        val pvtArgs = PvtArgs.fromBundle(intentExtras)
 
-        Log.d(TAG, "args: $pvtArgs")
+        viewModelFactory = PvtViewModelFactory(pvtArgs)
+        viewModel = ViewModelProvider(this, viewModelFactory).get(PvtViewModel::class.java)
 
-        pvt = Pvt(pvtArgs)
-
-        pvt?.setListener(
-            onStartCountdown = {
-                binding.textViewCountdown.visibility = View.VISIBLE
-                binding.textViewMessage.text = getString(R.string.ready_message)
-                binding.textViewMessage.visibility = View.VISIBLE
-                binding.viewStimulus.visibility = View.GONE
-            },
-            onCountdownUpdate = { updateCountdown(it) },
-            onFinishCountdown = {
-                binding.textViewCountdown.visibility = View.GONE
-                binding.textViewMessage.visibility = View.GONE
-                binding.textViewMessage.text = ""
-            },
-            onIntervalShowing = {
-                binding.textViewMessage.text = ""
-                binding.textViewMessage.visibility = View.GONE
-            },
-            onStimulusShowing = {
-                binding.viewStimulus.visibility = View.VISIBLE
-                binding.textViewMessage.visibility = View.VISIBLE
-            },
-            onStimulusHidden = {
-                binding.viewStimulus.visibility = View.GONE
-                binding.textViewMessage.visibility = View.GONE
-            },
-            onReactionDelayUpdate = {
-                updateReactionDelay(it)
-            },
-            onInvalidReaction = {
-                binding.textViewMessage.text = getString(R.string.invalid_reaction)
-                binding.textViewMessage.visibility = View.VISIBLE
-            },
-            onCompleteTest = {
-                binding.textViewMessage.visibility = View.VISIBLE
-                binding.textViewMessage.text = getString(R.string.test_complete)
-            },
-            onResults = {
-                val returnIntent = Intent()
-                returnIntent.putExtra(PVT_RESULTS_KEY, it)
-                setResult(RESULT_OK, returnIntent)
-                finish()
-            },
-        )
     }
 
     override fun onTouchEvent(event: MotionEvent?): Boolean {
         super.onTouchEvent(event)
 
-        if (event?.action == MotionEvent.ACTION_DOWN) {
-            pvt?.handleActionDownTouchEvent()
-        }
-
-        return true
+        return viewModel.handleOnTouchEvent(event)
     }
 
     private fun updateCountdown(millisElapsed: Long) {
@@ -89,52 +51,6 @@ class PvtActivity : AppCompatActivity() {
 
     private fun updateReactionDelay(millisElapsed: Long) {
         binding.textViewMessage.text = getString(R.string.reaction_delay, millisElapsed)
-    }
-
-    private fun getPvtArgsFromIntentExtras(bundle: Bundle?): Pvt.Args {
-        if (bundle == null) return Pvt.Args.default()
-
-        val keySet = bundle.keySet()
-
-        var stimulusCount: Int? = null
-        var minInterval: Long? = null
-        var maxInterval: Long? = null
-        var countDownTime: Long? = null
-        var stimulusTimeout: Long? = null
-        var postResponseDelay: Long? = null
-
-        if (keySet.contains(STIMULUS_COUNT)) {
-            stimulusCount = bundle.getInt(STIMULUS_COUNT)
-        }
-
-        if (keySet.contains(MIN_INTERVAL)) {
-            minInterval = bundle.getLong(MIN_INTERVAL)
-        }
-
-        if (keySet.contains(MAX_INTERVAL)) {
-            maxInterval = bundle.getLong(MAX_INTERVAL)
-        }
-
-        if (keySet.contains(COUNTDOWN_TIME)) {
-            countDownTime = bundle.getLong(COUNTDOWN_TIME)
-        }
-
-        if (keySet.contains(STIMULUS_TIMEOUT)) {
-            stimulusTimeout = bundle.getLong(STIMULUS_TIMEOUT)
-        }
-
-        if (keySet.contains(POST_RESPONSE_DELAY)) {
-            postResponseDelay = bundle.getLong(POST_RESPONSE_DELAY)
-        }
-
-        return Pvt.Args(
-            stimulusCount, minInterval, maxInterval, countDownTime, stimulusTimeout, postResponseDelay
-        )
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        pvt?.cancel()
     }
 
     class Builder {
@@ -204,12 +120,5 @@ class PvtActivity : AppCompatActivity() {
 
     companion object {
         private const val TAG = "PvtNew"
-        const val PVT_RESULTS_KEY = "pvtResultsKey"
-        const val STIMULUS_COUNT = "stimulusCount"
-        const val MIN_INTERVAL = "minInterval"
-        const val MAX_INTERVAL = "maxInterval"
-        const val COUNTDOWN_TIME = "countdownTime"
-        const val STIMULUS_TIMEOUT = "stimulusTimeout"
-        const val POST_RESPONSE_DELAY = "postResponseDelay"
     }
 }
