+++ AndroidPvtLib/src/main/java/rs/arthu/androidpvt/lib/PvtViewModel.kt
@@ -1,20 +1,32 @@
 package rs.arthu.androidpvt.lib;
 
+import android.os.Bundle
+import android.util.Log
+import android.view.MotionEvent
 import androidx.lifecycle.ViewModel
 
-class PvtViewModel : ViewModel() {
+internal class PvtViewModel(private val args: PvtArgs) : ViewModel(){
 
-    private lateinit var pvt: Pvt
+    private var pvt: Pvt
 
     init {
+        Log.d("PvtViewModel", "viewmodel created")
 
+        pvt = Pvt()
+    }
 
+    internal fun handleOnTouchEvent(event: MotionEvent?): Boolean {
+        if (event?.action == MotionEvent.ACTION_DOWN) {
+            pvt.handleActionDownTouchEvent()
+        }
+
+        return true
     }
 
     override fun onCleared() {
         super.onCleared()
-
         pvt.cancel()
+        Log.d("PvtViewModel", "oncleared")
     }
 
 }
